TopicListScreen [nfc]: Simplify a callback
The TopicList callback is passed the stream name as well as the topic,
so we can use the stream name from there, rather than mixing the
stream name from our props with the topic from the callback parameters.

That's good because the two parameters to the callback are clearly
meant to refer to the same conversation.  It happens that the existing
code works correctly too, because the stream from the props is always
the same stream that the listed topics are all from and the same
stream that the callback is called for... but that's a less local fact
that requires reading more code to see.  Better to do the thing that's
locally clearly right.

As a bonus, there's one less dependency for useCallback.  And we fix
the puzzling parameter name `streamObj` ("object"?) which had type
`string`, which is definitely not an object.  That bit dates to when
this component was first introduced, in 8ea968438.